Teach write_rss to retrieve content-length for each podcast entry
Each "enclosure" entry in an rss document has a "length" field which may
be filled in with the length in bytes of the url content.  If it's not
known then 0 should be supplied, which is what we've been doing.  Let's
teach write_rss how to determine the content length of each podcast and
fill in this field correctly when get_content_length keyword is True.